requirements.txt
flask
paho-mqtt
stmpy
requests
sense-hat
stmpy